+++ api/admin.py
@@ -17,11 +17,14 @@ from django.contrib.auth.admin import UserAdmin
 from django.contrib.auth.models import User
 from rest_framework.authtoken.admin import TokenAdmin
 from rest_framework.authtoken.models import Token
-from django.http import HttpResponse
+from django.http import HttpResponse, HttpResponseRedirect
 from .forms import ActionForm
 from reversion.admin import VersionAdmin
 from reversion.models import Revision, Version
 from reversion_compare.admin import CompareVersionAdmin
+
+from api.management.commands.index_and_notify import Command as Notify
+from notifications.models import RecordType, SubscriptionType
 
 class GoUserAdmin(UserAdmin):
     list_filter = (
@@ -182,6 +185,8 @@ class EventAdmin(CompareVersionAdmin, RegionRestrictedAdmin):
         return mark_safe('<span class="errors">No related appeals</span>')
     appeals.short_description = 'Appeals'
 
+    # To add the 'Notify subscribers now' button
+    change_form_template = "admin/emergency_changeform.html"
     # Overwriting readonly fields for Edit mode
     def changeform_view(self, request, *args, **kwargs):
         self.readonly_fields = list(self.readonly_fields)
@@ -190,6 +195,18 @@ class EventAdmin(CompareVersionAdmin, RegionRestrictedAdmin):
                 self.readonly_fields.append('parent_event')
 
         return super(EventAdmin, self).changeform_view(request, *args, **kwargs)
+
+    # Evaluate if the regular 'Save' or 'Notify subscribers now' button was pushed
+    def response_change(self, request, obj):
+        if "_notify-subscribers" in request.POST and request.user.is_superuser:
+            notif_class = Notify()
+            try:
+                notif_class.notify(records=[obj], rtype=RecordType.FOLLOWED_EVENT, stype=SubscriptionType.NEW, uid=request.user.id)
+                self.message_user(request, "Successfully notified subscribers.")
+            except:
+                self.message_user(request, "Could not notify subscribers.", level=messages.ERROR)
+            return HttpResponseRedirect(".")
+        return super().response_change(request, obj)
 
     def field_reports(self, instance):
         if getattr(instance, 'field_reports').exists():
@@ -247,7 +264,12 @@ class FieldReportAdmin(CompareVersionAdmin, RegionRestrictedAdmin):
     list_select_related = ('event',)
     search_fields = ('countries__name', 'regions__name', 'summary',)
     autocomplete_fields = ('event', 'countries', 'districts',)
-    readonly_fields = ('report_date', 'created_at', 'updated_at',)
+    readonly_fields = (
+        'report_date', 'created_at', 'updated_at',
+        'health_min_cases', 'health_min_suspected_cases', 'health_min_probable_cases', 'health_min_confirmed_cases', 'health_min_num_dead',
+        'who_cases', 'who_suspected_cases', 'who_probable_cases', 'who_confirmed_cases', 'who_num_dead',
+        'other_cases', 'other_suspected_cases', 'other_probable_cases', 'other_confirmed_cases'
+    )
     list_filter = [MembershipFilter,]
     actions = ['create_events', 'export_field_reports', ]
 
@@ -486,9 +508,25 @@ class SituationReportTypeAdmin(CompareVersionAdmin):
     search_fields = ('type',)
 
 class CronJobAdmin(CompareVersionAdmin):
+    list_display = ('name', 'created_at', 'num_result', 'status')
     search_fields = ('name', 'created_at',)
     readonly_fields = ('created_at',)
     list_filter = ('status', 'name')
+    readonly_fields = ('message_display',)
+
+    def message_display(self, obj):
+        style_class = {
+            models.CronJobStatus.WARNED: 'warning',
+            models.CronJobStatus.ERRONEOUS: 'error',
+        }.get(obj.status, 'success')
+        if obj.message:
+            return mark_safe(
+                f'''
+                <ul class="messagelist" style="margin-left: 0px;">
+                    <li class="{style_class}"><pre>{obj.message}</pre></li>
+                </ul>
+                '''
+            )
 
 
 class EmergencyOperationsDatasetAdmin(CompareVersionAdmin):