+++ databank/management/commands/sources/FDRS.py
@@ -3,7 +3,6 @@ import requests
 from django.conf import settings
 
 from api.utils import base64_encode
-from api.models import CronJob, CronJobStatus
 from databank.models import CountryOverview as CO
 from .utils import catch_error
 
@@ -44,11 +43,7 @@ FDRS_DATA_API_ENDPOINT = 'https://data-api.ifrc.org/api/Data?indicator=' + ','.j
 @catch_error('Error occured while fetching from FDRS API, Please make sure valid FDRS_CREDENTIAL is provided')
 def prefetch():
     fdrs_entities = requests.get(FDRS_NS_API_ENDPOINT, headers=FDRS_HEADERS)
-
-    if fdrs_entities.status_code != 200:
-        body = { "name": "FDRS", "message": "Error querying FDRS NS API feed at " + FDRS_NS_API_ENDPOINT, "status": CronJobStatus.ERRONEOUS } # not every case is catched here, e.g. if the base URL is wrong...
-        CronJob.sync_cron(body)
-        return {}
+    fdrs_entities.raise_for_status()
     fdrs_entities = fdrs_entities.json()
 
     ns_iso_map = {
@@ -56,9 +51,6 @@ def prefetch():
         ns['KPI_DON_code']: ns['iso_2']
         for ns in fdrs_entities
     }
-
-    body = { "name": "FDRS", "message": "Done querying FDRS NS API feed at " + FDRS_NS_API_ENDPOINT, "num_result": len(ns_iso_map), "status": CronJobStatus.SUCCESSFUL }
-    CronJob.sync_cron(body)
 
     return {
         # KEY <ISO2>-<Indicator_ID>: {year: '', value: ''}
@@ -69,7 +61,7 @@ def prefetch():
         )
         for indicator_data in requests.get(FDRS_DATA_API_ENDPOINT, headers=FDRS_HEADERS).json()['data']
         for ns_data in indicator_data['data']
-    }
+    }, len(ns_iso_map), FDRS_DATA_API_ENDPOINT
 
 
 @catch_error()