+++ databank/management/commands/sources/WB.py
@@ -2,7 +2,7 @@ import requests
 import datetime
 import logging
 
-from api.models import District, CronJob, CronJobStatus
+from api.models import District
 
 from .utils import catch_error, get_country_by_iso3
 
@@ -27,10 +27,7 @@ def prefetch():
             'per_page': 5000 - 1,  # WD throws error on 5000
             'page': page,
         })
-        if rs.status_code != 200:
-            body = { "name": "WB", "message": "Error querying WorldBank feed at " + url, "status": CronJobStatus.ERRONEOUS } # not every case is catched here, e.g. if the base URL is wrong...
-            CronJob.sync_cron(body)
-            return data
+        rs.raise_for_status()
         rs = rs.json()
 
         for pop_data in rs[1]:
@@ -58,9 +55,7 @@ def prefetch():
         if page >= rs[0]['pages']:
             break
         page += 1
-    body = { "name": "WB", "message": "Done querying WorldBank feed at " + url, "num_result": len(data), "status": CronJobStatus.SUCCESSFUL }
-    CronJob.sync_cron(body)
-    return data
+    return data, len(data), API_ENDPOINT
 
 
 @catch_error()